Centralised makefile for benchmarks
This is the starting point for some common infrastructure to compile
and run all benchmarks. It recursively calls make and expects that
each directory with a benchmark program contains a makefile with
the following targets:

<default> compile the benchmark
   (avoid recompilation by defining correct dependencies..)
run_X  run the benchmark with data set X
   (where X is in {small,medium,large})
clean

Further centralised targets are defined...
documentation should go into the Makefile itself one day.